refactor cpp compiler runtime
